experiment-control/requirements.txt
paramiko
socketIO_client_nexus
watchdog
paho-mqtt
argparse
pytest